src/fastapi-app/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
aiohttp==3.9.1
pyjwt==2.8.0